Extract base class for page cursors klasses
